Waypoint Following
The UnitAI scipt now has the ability to follow predetermined waypoints. This works for climbing robots on the road and also for parrot drones. Because the Move/TrussMove command types always find desired headingbased on the XZ plane, the climbing robot cannot yet follow move commands while on a truss. I just need to correct TrussMove to find desired heading based on the plane of movement of the entity instead of always using the XZ plane and this will be resolved.