Camera viewports now generated at runtime
The UIMgr now Instantiates the camera view ports for each entity at runtime. This includes the CameraEntity since it is a StacsEntity, which may be undesirable. Also, this UI now would be improved with a scollview for the containing panel since the viewports are compressed horizontally because there are too many entities.